polish(pu): refactor
